Fix Floating Islands order
Resolves #729.
