+++ TRRandomizerCore/Randomizers/TR2/Classic/TR2SecretRandomizer.cs
@@ -1,4 +1,5 @@
 ﻿using TRGE.Core;
+using TRLevelControl.Helpers;
 using TRLevelControl.Model;
 using TRRandomizerCore.Helpers;
 using TRRandomizerCore.Levels;
@@ -8,6 +9,13 @@ namespace TRRandomizerCore.Randomizers;
 
 public class TR2SecretRandomizer : BaseTR2Randomizer
 {
+    private static readonly List<string> _textureFixLevels = new()
+    {
+        TR2LevelNames.FLOATER,
+        TR2LevelNames.LAIR,
+        TR2LevelNames.HOME
+    };
+
     public IMirrorControl Mirrorer { get; set; }
     public ItemFactory<TR2Entity> ItemFactory { get; set; }
 
@@ -34,6 +42,8 @@ public class TR2SecretRandomizer : BaseTR2Randomizer
                 List<Location> pickedLocations = allocator.RandomizeSecrets(_levelInstance.Name, _levelInstance.Data);
                 AddDamageControl(_levelInstance, pickedLocations);
             }
+
+            FixSecretTextures(_levelInstance.Name, _levelInstance.Data);
 
             SaveLevelInstance();
             if (!TriggerProgress())
@@ -78,5 +88,17 @@ public class TR2SecretRandomizer : BaseTR2Randomizer
 
         level.Script.AddStartInventoryItem(ItemUtilities.ConvertToScriptItem(TR2Type.LargeMed_S_P), hardDamageCount);
         level.Script.AddStartInventoryItem(ItemUtilities.ConvertToScriptItem(TR2Type.SmallMed_S_P), easyDamageCount);
+    }
+
+    private static void FixSecretTextures(string levelName, TR2Level level)
+    {
+        if (!_textureFixLevels.Contains(levelName))
+        {
+            return;
+        }
+
+        // Swap Stone and Jade textures - OG has them the wrong way around.
+        (level.Sprites[TR2Type.JadeSecret_S_P].Textures, level.Sprites[TR2Type.StoneSecret_S_P].Textures)
+            = (level.Sprites[TR2Type.StoneSecret_S_P].Textures, level.Sprites[TR2Type.JadeSecret_S_P].Textures);
     }
 }
