+++ TRRandomizerCore/Randomizers/TR2/Shared/TR2SecretAllocator.cs
@@ -11,13 +11,6 @@ public class TR2SecretAllocator : ISecretRandomizer
 {
     private static readonly int _levelSecretCount = 3;
 
-    private static readonly List<string> _textureFixLevels = new()
-    {
-        TR2LevelNames.FLOATER,
-        TR2LevelNames.LAIR,
-        TR2LevelNames.HOME
-    };
-
     private readonly Dictionary<string, List<Location>> _locations;
     private readonly LocationPicker _routePicker;
     private SecretPicker<TR2Entity> _secretPicker;
@@ -75,8 +68,6 @@ public class TR2SecretAllocator : ISecretRandomizer
             int zone = zones.FindIndex(z => z.Contains(location.Room));
             PlaceSecret(entity, secretTypes[zone], location);
         }
-
-        FixSecretTextures(levelName, level);
     }
 
     public List<Location> RandomizeSecrets(string levelName, TR2Level level)
@@ -117,9 +108,6 @@ public class TR2SecretAllocator : ISecretRandomizer
         }
 
         _secretPicker.FinaliseSecretPool(pickedLocations, levelName, itemIndex => GetDependentLockedItems(level, itemIndex));
-
-        FixSecretTextures(levelName, level);
-
         return pickedLocations;
     }
 
@@ -148,17 +136,5 @@ public class TR2SecretAllocator : ISecretRandomizer
         entity.Intensity1 = -1;
         entity.Intensity2 = -1;
         entity.Flags = 0;
-    }
-
-    private static void FixSecretTextures(string levelName, TR2Level level)
-    {
-        if (!_textureFixLevels.Contains(levelName))
-        {
-            return;
-        }
-
-        // Swap Stone and Jade textures - OG has them the wrong way around.
-        (level.Sprites[TR2Type.JadeSecret_S_P].Textures, level.Sprites[TR2Type.StoneSecret_S_P].Textures)
-            = (level.Sprites[TR2Type.StoneSecret_S_P].Textures, level.Sprites[TR2Type.JadeSecret_S_P].Textures);
     }
 }
